Refactor to use single activity and multiple composables using Jetpack Compose Navigation.
