+++ app/src/main/java/com/davidread/habittracker/common/model/Screen.kt
@@ -1,0 +1,7 @@
+package com.davidread.habittracker.common.model
+
+sealed class Screen(val route: String) {
+    object Login : Screen("login")
+    object Signup : Screen("signup")
+    object HabitList : Screen("habit_list")
+}
