+++ app/src/main/java/com/davidread/habittracker/common/ui/activity/MainActivity.kt
@@ -1,0 +1,21 @@
+package com.davidread.habittracker.common.ui.activity
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import com.davidread.habittracker.common.ui.composable.HabitTrackerApp
+import com.davidread.habittracker.common.ui.theme.HabitTrackerTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            HabitTrackerTheme {
+                HabitTrackerApp()
+            }
+        }
+    }
+}
